finishing-departments

+++ src/Components/UI/Tables/DepartmentTable.js
@@ -1,27 +1,42 @@
-import { useEffect, useState } from 'react';
 import Table from 'react-bootstrap/Table';
 import DepartmentDataService from '../../../Services/DepartmentDataService';
-import { Link } from 'react-router-dom';
+import Swal from "sweetalert2";
 
-const DepartmentsTable = () => {
-    const [departments, setDepartments] = useState([]);
-
-    const fetchData = async () => {
-        const departmentsData = await DepartmentDataService.getAll()
-        setDepartments(departmentsData.data);
+const DepartmentsTable = ({ handleOpenForm, departments, fetchData }) => {
+    const deleteHandler = async (id, event) => {
+        Swal.fire({
+            title: 'Do you want to delete this department?',
+            showDenyButton: true,
+            confirmButtonText: 'Yes',
+            denyButtonText: 'No',
+            confirmButtonColor: '#007bff',
+            customClass: {
+                actions: 'my-actions',
+                cancelButton: 'order-1 right-gap',
+                confirmButton: 'order-2',
+                denyButton: 'order-3',
+            },
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                Swal.fire({
+                    title: "Deparetment was deleted!",
+                    icon: "success",
+                    toast: true,
+                    timer: 1000,
+                    position: 'top-right',
+                    timerProgressBar: true,
+                    showConfirmButton: false,
+                });
+                await DepartmentDataService.deleteById(id);
+                await fetchData();
+            } else if (result.isDenied) {
+                return;
+            }
+        })
     }
-
-    useEffect(() => {
-        fetchData().catch(console.error);
-    }, []);
 
     const addRows = () => {
         return departments.map(d => createRow(d));
-    }
-
-    const deleteHandler = async (id, event) => {
-        await DepartmentDataService.deleteById(id);
-        await fetchData();
     }
  
     const createRow = department => {
@@ -30,18 +45,13 @@ const DepartmentsTable = () => {
                 <td>{department.name}</td>
                 <td>{department.address}</td>
                 <td>
-                    <Link className='btn btn-outline-secondary' to={`edit?id=${department.id}`}>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pen" viewBox="0 0 16 16">
-                            <path d="m13.498.795.149-.149a1.207 1.207 0 1 1 1.707 1.708l-.149.148a1.5 1.5 0 0 1-.059 2.059L4.854 14.854a.5.5 0 0 1-.233.131l-4 1a.5.5 0 0 1-.606-.606l1-4a.5.5 0 0 1 .131-.232l9.642-9.642a.5.5 0 0 0-.642.056L6.854 4.854a.5.5 0 1 1-.708-.708L9.44.854A1.5 1.5 0 0 1 11.5.796a1.5 1.5 0 0 1 1.998-.001m-.644.766a.5.5 0 0 0-.707 0L1.95 11.756l-.764 3.057 3.057-.764L14.44 3.854a.5.5 0 0 0 0-.708z"/>
-                        </svg>
-                    </Link>
+                    <button className='btn btn-outline-primary' onClick={ event => { handleOpenForm(department.id, event) } }>
+                        <i class="bi bi-pen"></i>
+                    </button>
                 </td>
                 <td>
-                    <button className='btn btn-outline-danger' onClick={(event) => {deleteHandler(department.id, event)}}>
-                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash" viewBox="0 0 16 16">
-                            <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z"></path>
-                            <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z"></path>
-                        </svg>
+                    <button className='btn btn-outline-danger' onClick={ event => { deleteHandler(department.id, event) } }>
+                        <i class="bi bi-trash"></i>
                     </button>
                 </td>
             </tr>
